Added a new class for easier plotting and added preoption feature
Added a class to allow for easy plotting. Attempting to abstract some of the boilerplate away from the user.
Added an option to add a string before the file in filename plots. Ie 'plot for[i=1:2] filename.dat
Added two methods to create appropriate .dat files for each EzPlot method